Develop ey

+++ student-welfare/src/main/java/in/cdac/university/studentWelfare/bean/SavitbpSchemeApplTrackerdtlBean.java
@@ -22,7 +22,7 @@ public class SavitbpSchemeApplTrackerdtlBean {
 	
 	private Long 	unumStudentId;
 	
-	private Long 	unumSno;
+	private Integer	unumSno;
 	
 	private Long 	unumApplLevelid;
